refactor: Fixed relative imports, added all stages to makefile, tested
