+++ articles/ml_system_design/real_time_news_search_with_upstash_kafka_and_vector_db/src/__init__.py
@@ -1,0 +1,5 @@
+import os
+import sys
+
+root = os.path.abspath(os.path.dirname(__file__))
+sys.path.insert(0, root)
