+++ articles/ml_system_design/real_time_news_search_with_upstash_kafka_and_vector_db/src/embeddings.py
@@ -12,8 +12,8 @@ from typing import Optional, Union
 import numpy as np
 from transformers import AutoModel, AutoTokenizer
 
-from .settings import settings
-from .logger import get_logger
+from settings import settings
+from logger import get_logger
 
 logger = get_logger(__name__)
