+++ articles/ml_system_design/real_time_news_search_with_upstash_kafka_and_vector_db/src/settings.py
@@ -4,7 +4,6 @@ from pydantic_settings import SettingsConfigDict, BaseSettings
 import os
 
 dir_path = os.path.dirname(os.path.realpath(__file__))
-print(dir_path)
 
 
 class AppSettings(BaseSettings):
