+++ articles/ml_system_design/real_time_news_search_with_upstash_kafka_and_vector_db/src/start.py
@@ -4,7 +4,7 @@
 """
 
 from bytewax.testing import run_main
-from .flow import build as build_flow
+from flow import build as build_flow
 
 flow = build_flow()
