+++ articles/ml_system_design/real_time_news_search_with_upstash_kafka_and_vector_db/src/tools.py
@@ -7,8 +7,8 @@ from newsapi import NewsApiClient
 from newsdataapi import NewsDataApiClient
 from pydantic import ValidationError
 
-from .models import NewsAPIModel, NewsDataIOModel
-from .settings import settings
+from models import NewsAPIModel, NewsDataIOModel
+from settings import settings
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
