+++ articles/ml_system_design/real_time_news_search_with_upstash_kafka_and_vector_db/src/vector.py
@@ -2,9 +2,9 @@ from typing import Optional, List
 
 from bytewax.outputs import DynamicSink, StatelessSinkPartition
 from upstash_vector import Index, Vector
-from .models import EmbeddedDocument
-from .settings import settings
-from .logger import get_logger
+from models import EmbeddedDocument
+from settings import settings
+from logger import get_logger
 
 
 logger = get_logger(__name__)
