+++ articles/ml_system_design/real_time_news_search_with_upstash_kafka_and_vector_db/ui.py
@@ -10,13 +10,12 @@
 """
 
 from io import BytesIO
-
 import requests
 import streamlit as st
 from PIL import Image
-from embeddings import TextEmbedder
-from settings import settings
-from cleaners import clean_full
+from src.embeddings import TextEmbedder
+from src.settings import settings
+from src.cleaners import clean_full
 from upstash_vector import Index
 
 v_index = Index(url=settings.UPSTASH_VECTOR_ENDPOINT, token=settings.UPSTASH_VECTOR_KEY)
